// src/types/mension-response.type.ts
export type MensionResponse = {
  createdAt: string; // "2026-02-02T16:11.06"
  updatedAt: string; // "2026-02-02T16:11.06"
  id: string; // "001a0607-5d83-4862-b4f4-8916d2b7676c"
  time: string; // "2026-02-02"
  postId: string; // "3611948387932288145"
  postType: string; // "clips"
  shortCode: string; // "DIgN-wBv7SR"
  postUrl: string; // "https://www.instagram.com/p/DIgN-wBv7SR/"
  caption: string; // "F45 합정 X 사계단백연구소..."
  likesCount: number; // 49
  commentsCount: number; // 1
  videoPlayCount: number; // 3162
  igPlayCount: number; // 3162
  reshareCount: number; // 0
  videoDuration: number; // 105.76
  postedAt: string; // "2025-04-16T10:08:06"
  ownerId: string; // "62526264689"
  ownerUsername: string; // "f45_hapjeong"
  ownerFullName: string; // "F45 합정"
  ownerProfilePicUrl: string | null; // null
  displayUrl: string; // "https://scontent-lax7-1.cdninstagram.com/..."
  videoUrl: string; // "https://scontent-lax3-1.cdninstagram.com/..."
  images: string; // "[]"
  hashtags: string; // "[\"f45\",\"f45합정\",...]"
  mentions: string; // "[\"allseasons_protein_lab\",\"f45_hapjeong\"]"
  taggedUsers: string; // "[{\"id\",\"username\",\"fullname\",\"profile_pic_url\",\"_private\",\"_verified\"}]"
  musicInfo: string; // "{\"artist_name\":\"f45_hapjeong\",\"song_name\":\"Original audio\",...}"
  coauthorProducers: string; // "[{\"full_name\":\"사계단백연구소\",\"username\":\"allseasons_protein_lab\"}]"
  childPosts: string; // "[]"
};